+++ crates/core/src/loader.rs
@@ -59,17 +59,8 @@ pub fn locate_tool(id: &Id, proto: &ProtoEnvironment) -> Result<PluginLocator, P
     }
 
     // Search in registries
-    if locator.is_none()
-        && let Some(default_registry) = &config.settings.default_registry
-        && let Some(registry) = config
-            .settings
-            .registries
-            .iter()
-            .find(|reg| &reg.registry == default_registry)
-    {
-        if let Ok(maybe_locator) =
-            PluginLocator::try_from(format!("registry://{}", registry.get_reference(id)))
-        {
+    if locator.is_none() && !config.settings.registries.is_empty() {
+        if let Ok(maybe_locator) = PluginLocator::try_from(format!("registry://{id}")) {
             debug!(
                 plugin = maybe_locator.to_string(),
                 "Using a registry plugin"